+++ src/main/java/faang/school/projectservice/service/internship/InternshipService.java
@@ -1,9 +1,9 @@
 package faang.school.projectservice.service.internship;
 
-import faang.school.projectservice.dto.filter.InternshipFilterDto;
+import faang.school.projectservice.dto.filter.internship.InternshipFilterDto;
 import faang.school.projectservice.dto.internship.InternshipDto;
 import faang.school.projectservice.exception.DataValidationException;
-import faang.school.projectservice.filter.internship.Filter;
+import faang.school.projectservice.filter.Filter;
 import faang.school.projectservice.mapper.internship.InternshipMapper;
 import faang.school.projectservice.model.Internship;
 import faang.school.projectservice.model.InternshipStatus;
