+++ src/test/java/faang/school/projectservice/controller/internship/InternshipControllerTest.java
@@ -1,6 +1,6 @@
 package faang.school.projectservice.controller.internship;
 
-import faang.school.projectservice.dto.filter.InternshipFilterDto;
+import faang.school.projectservice.dto.filter.internship.InternshipFilterDto;
 import faang.school.projectservice.dto.internship.InternshipDto;
 import faang.school.projectservice.service.internship.InternshipService;
 import org.junit.jupiter.api.DisplayName;
