+++ src/test/java/faang/school/projectservice/filter/internship/InternshipStatusFilterTest.java
@@ -1,6 +1,6 @@
 package faang.school.projectservice.filter.internship;
 
-import faang.school.projectservice.dto.filter.InternshipFilterDto;
+import faang.school.projectservice.dto.filter.internship.InternshipFilterDto;
 import faang.school.projectservice.model.Internship;
 import faang.school.projectservice.model.InternshipStatus;
 import org.junit.jupiter.api.DisplayName;
